refactor: split up tsconfig per context
src, scripts, and tests use separate configs now
