+++ tsconfig.scripts.json
@@ -1,0 +1,9 @@
+{
+	"extends": "./tsconfig.base.json",
+	"compilerOptions": {
+		"typeRoots": ["./types", "./node_modules/@types"],
+		"types": ["node", "babel__core"]
+	},
+	"include": ["*.js", "package.json", "scripts/**/*"],
+	"exclude": ["**/coverage"]
+}
