+++ tsconfig.src.json
@@ -1,0 +1,9 @@
+{
+	"extends": "./tsconfig.base.json",
+	"compilerOptions": {
+		"typeRoots": ["./types", "./node_modules/@types"],
+		"types": ["node", "semver", "vscode", "path-is-inside"]
+	},
+	"include": ["src/**/*", "types/*.d.ts"],
+	"exclude": ["**/coverage", "**/__tests__", "**/__mocks__", "types/tests.d.ts"]
+}
